% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{executeTreatmentPatterns}
\alias{executeTreatmentPatterns}
\title{This is the main function which runs all different parts of the treatment pathways analysis.}
\usage{
executeTreatmentPatterns(
  connection = NULL,
  connectionDetails = NULL,
  cdmDatabaseSchema,
  cohortDatabaseSchema = cdmDatabaseSchema,
  cohortTable = "cohort",
  outputFolder,
  databaseName = "Unknown",
  runCreateCohorts = TRUE,
  runCohortCharacterization = FALSE,
  runConstructPathways = TRUE,
  runGenerateResults = TRUE,
  study_settings
)
}
\arguments{
\item{connection}{Connection to database server.}

\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{cohortDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{cohortTable}{The name of the table that will be created in the work database schema.
This table will hold the target and event cohorts used in this
study.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/).}

\item{databaseName}{Name of the database that will appear in the results.}

\item{runCreateCohorts}{Setting to run part of the analysis where cohorts are created.}

\item{runCohortCharacterization}{Setting to run part of the analysis where characterization of target cohorts is done.}

\item{runConstructPathways}{Setting to run part of the analysis where treatment pathways are constructed.}

\item{runGenerateResults}{Setting to run part of the analysis where final result files and plots are generated.}

\item{study_settings}{Object that contains all study settings inputted by the user.}
}
\description{
This is the main function which runs all different parts of the treatment pathways analysis.
}
